Add claims and authorization

+++ resources/item.py
@@ -1,7 +1,7 @@
 from flask import current_app
 from flask.views import MethodView
 from flask_smorest import Blueprint, abort
-from flask_jwt_extended import jwt_required
+from flask_jwt_extended import jwt_required, get_jwt
 from sqlalchemy.exc import SQLAlchemyError, IntegrityError
 
 from db import db
@@ -14,12 +14,19 @@ blp = Blueprint('items', __name__, description='Operations on items')
 
 @blp.route('/item/<int:item_id>')
 class Item(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema)
     def get(self, item_id):
         item: ItemModel = ItemModel.query.get_or_404(item_id)
         return item
 
+    @jwt_required()
     def delete(self, item_id):
+        jwt = get_jwt()
+        current_app.logger.info(f'jwt: {jwt}')
+        if not jwt.get('is_admin'):
+            abort(401, message='Admin privilege required.')
+
         item: ItemModel = ItemModel.query.get_or_404(item_id)
         db.session.delete(item)
         db.session.commit()
@@ -30,6 +37,7 @@ class Item(MethodView):
 
         return {'message': 'Item deleted successfully.'}
 
+    @jwt_required()
     @blp.arguments(ItemUpdateSchema)
     @blp.response(200, ItemSchema)
     def put(self, item_data, item_id):
@@ -55,8 +63,14 @@ class Item(MethodView):
 
 @blp.route('/item')
 class ItemList(MethodView):
+    @jwt_required()
     @blp.response(200, ItemSchema(many=True))
     def get(self):
+        
+        jwt = get_jwt()
+        current_app.logger.info(f'jwt: {jwt}')
+        if not jwt.get('is_admin'):
+            abort(401, message='Admin privilege required.')
         return ItemModel.query.all()
 
     @jwt_required()
